pyproject.toml
[tool.poetry]
name = "pipit"
version = "0.1.0"
description = ""
authors = ["Matthew Rehberg"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.1"
flask-navigation = "^0.2.0"
networkx = "^3.2.1"
pyenchant = "^3.2.2"
pydictionary = "^2.0.1"
tzdata = "^2024.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"